[ENH] Create stats/utils.py to organize fetching functions
