+++ .coveragerc
@@ -38,6 +38,7 @@ omit =
 	pynets/stats/benchmarking.py
 	pynets/stats/embeddings.py
 	pynets/stats/prediction.py
+        pynets/stats/utils.py
 	pynets/registration/register.py
 	pynets/core/interfaces.py
 	pynets/core/cloud_utils.py
