+++ pynets/runconfig.yaml
@@ -138,7 +138,7 @@ available_models:
       - 'sfm'
 gradients:
     n_components:
-        - null
+        - 1
 #sentiment_labeling:
 #    liwc_file:
 #        - '/Users/derekpisner/Documents/LIWC2015_English.dic' # Must be purchased and licensed. Visit https://liwc.wpengine.com/ for more information.
